+++ src/main/java/open/commons/spring/web/beans/authority/FieldAccessAuthorityDecision.java
@@ -18,7 +18,7 @@
  *
  * This file is generated under this project, "open-commons-spring-web".
  *
- * Date  : 2025. 5. 19. 오후 5:39:00
+ * Date  : 2025. 10. 1. 오후 5:51:58
  *
  * Author: parkjunhong77@gmail.com
  * 
@@ -26,43 +26,50 @@
 
 package open.commons.spring.web.beans.authority;
 
-import javax.validation.constraints.NotEmpty;
-
-import open.commons.core.Result;
-import open.commons.core.TwoValueObject;
-
 /**
- * 클래스 필드 접근권한 제공 서비스.
  * 
- * @since 2025. 5. 19.
+ * @since 2025. 10. 1.
  * @version 0.8.0
  * @author parkjunhong77@gmail.com
  */
-public interface IFieldAccessAuthorityProvider extends IResourceAccessAuthorityProvider {
+public class FieldAccessAuthorityDecision {
+
+    /** 허용 여부 */
+    public final boolean accessible;
+    /** 데이터 처리 방식 */
+    public final int handleType;
+    /**
+     * 데이터 처리 빈<br>
+     * {@link IUnauthorizedFieldHandler} 를 구현한 객체여야 함.
+     */
+    public final String handleBean;
 
     /**
-     * 클래스의 특정 필드에 대한 처리방식을 제공합니다. <br>
+     * 
+     * <br>
      * 
      * <pre>
      * [개정이력]
      *      날짜    	| 작성자	|	내용
      * ------------------------------------------
-     * 2025. 5. 27.		박준홍			최초 작성
+     * 2025. 10. 1.		박준홍			최초 작성
      * </pre>
      *
-     * @param type
-     *            데이터 유형<br>
-     *            {@link Class#toString()} 값이 전달됨.
-     * @param fieldName
-     *            필드 이름
-     * @return
-     *         <li>{@link TwoValueObject#first}: 데이터 제공 허용 여부
-     *         <li>{@link TwoValueObject#second}: 허용되지 않은 경우 처리 방식.
+     * @param accessible
+     *            허용 여부
+     * @param handleType
+     *            데이터 처리 방식
+     * @param handleBean
+     *            데이터 처리 빈<br>
+     *            {@link IUnauthorizedFieldHandler} 를 구현한 객체여야 함.
      *
-     * @since 2025. 5. 26.
+     * @since 2025. 10. 1.
      * @version 0.8.0
-     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     * @author parkjunhong77@gmail.com
      */
-    Result<FieldAccessAuthorityDecision> isAllowed(@NotEmpty String type, @NotEmpty String fieldName);
-
+    public FieldAccessAuthorityDecision(boolean accessible, int handleType, String handleBean) {
+        this.accessible = accessible;
+        this.handleType = handleType;
+        this.handleBean = handleBean;
+    }
 }