+++ src/main/java/open/commons/spring/web/beans/authority/internal/ForcedUnintelligibleJudge.java
@@ -31,9 +31,10 @@ import java.lang.reflect.Field;
 import javax.validation.constraints.NotEmpty;
 
 import open.commons.core.Result;
-import open.commons.core.TwoValueObject;
 import open.commons.core.utils.ExceptionUtils;
+import open.commons.spring.web.beans.authority.FieldAccessAuthorityDecision;
 import open.commons.spring.web.beans.authority.IFieldAccessAuthorityProvider;
+import open.commons.spring.web.beans.authority.internal.ForcedUnintelligibleHandler.ForcedUnintelligibleHandleType;
 import open.commons.spring.web.servlet.InternalServerException;
 
 /**
@@ -73,14 +74,14 @@ public class ForcedUnintelligibleJudge implements IFieldAccessAuthorityProvider 
      * @see open.commons.spring.web.beans.authority.IFieldAccessAuthorityProvider#isAllowed(java.lang.String,
      *      java.lang.String)
      */
-    @SuppressWarnings("unchecked")
     @Override
-    public Result<TwoValueObject<Boolean, Integer>> isAllowed(@NotEmpty String type, @NotEmpty String fieldName) {
+    public Result<FieldAccessAuthorityDecision> isAllowed(@NotEmpty String type, @NotEmpty String fieldName) {
         try {
             Class<?> typeClass = Class.forName(type);
             Field field = typeClass.getDeclaredField(fieldName);
             Class<?> fieldClass = field.getType();
-            return Result.success(new TwoValueObject(false, ForcedUnintelligibleHandleType.find(fieldClass)));
+
+            return Result.success(new FieldAccessAuthorityDecision(false, ForcedUnintelligibleHandleType.find(fieldClass), ForcedUnintelligibleHandler.BEAN_QUALIFIER));
         } catch (ClassNotFoundException | NoSuchFieldException | SecurityException e) {
             throw ExceptionUtils.newException(InternalServerException.class, e, "필드유형에 따른 처리 방식을 결정하는 도중에 오류가 발생하였습니다. 원인=%s", e.getMessage());
         }