+++ src/main/java/open/commons/spring/web/jackson/serialization/AuthorizedFieldDecisionUtil.java
@@ -32,10 +32,10 @@ import java.util.UUID;
 import org.springframework.core.annotation.AnnotationUtils;
 
 import open.commons.core.Result;
-import open.commons.core.TwoValueObject;
 import open.commons.core.utils.ExceptionUtils;
 import open.commons.spring.web.authority.AuthorizedField;
 import open.commons.spring.web.authority.metadata.AuthorizedFieldMetadata;
+import open.commons.spring.web.beans.authority.FieldAccessAuthorityDecision;
 import open.commons.spring.web.beans.authority.IAuthorizedResourcesMetadata;
 import open.commons.spring.web.beans.authority.IFieldAccessAuthorityProvider;
 import open.commons.spring.web.servlet.InternalServerException;
@@ -58,20 +58,21 @@ public class AuthorizedFieldDecisionUtil {
         return BeanUtils.isSimpleValueType(type) || UUID.class.equals(type);
     }
 
-    public static TwoValueObject<Boolean, Integer> resolve(Class<?> serializedType, AnnotatedField annotatedField, IFieldAccessAuthorityProvider fieldAccessor,
+    public static FieldAccessAuthorityDecision resolve(Class<?> serializedType, AnnotatedField annotatedField, IFieldAccessAuthorityProvider fieldAccessor,
             IAuthorizedResourcesMetadata metadata) {
 
         final Field field = annotatedField.getAnnotated();
         final Class<?> declaringClass = annotatedField.getDeclaringClass();
         final String fieldName = annotatedField.getName();
 
-        int handle = AuthorizedField.NO_ASSINGED_HANDLE_TYPE;
         boolean accessible = false;
+        int handleType = AuthorizedField.NO_ASSINGED_HANDLE_TYPE;
+        String handleBean = null;
 
         // 1) @AuthorizedField 우선
         AuthorizedField authField = AnnotationUtils.findAnnotation(field, AuthorizedField.class);
         if (authField != null) {
-            handle = authField.handleType();
+            handleType = authField.handleType();
         } else {
             // 2) 메타데이터 조회 (serializedType → declaringClass 순)
             AuthorizedFieldMetadata afm = metadata.getAuthorizedFieldMetadata(serializedType, fieldName);
@@ -79,28 +80,29 @@ public class AuthorizedFieldDecisionUtil {
                 afm = metadata.getAuthorizedFieldMetadata(declaringClass, fieldName);
             }
             if (afm != null) {
-                handle = afm.getHandleType();
+                handleType = afm.getHandleType();
             }
         }
 
         try {
             // 3) 아직 미지정이면 권한서비스로 접근/처리결정 조회
-            if (handle == AuthorizedField.NO_ASSINGED_HANDLE_TYPE) {
-                Result<TwoValueObject<Boolean, Integer>> res = fieldAccessor.isAllowed(declaringClass.getName(), fieldName);
-                if (res == null || res.isError() || res.getData() == null) {
-                    throw ExceptionUtils.newException(InternalServerException.class, "필드 접근권한 조회 실패. type=%s, field=%s, result=%s", declaringClass, fieldName, res);
+            if (handleType == AuthorizedField.NO_ASSINGED_HANDLE_TYPE) {
+                Result<FieldAccessAuthorityDecision> resultFieldAccessorDecision = fieldAccessor.isAllowed(declaringClass.getName(), fieldName);
+                if (resultFieldAccessorDecision == null || resultFieldAccessorDecision.isError() || resultFieldAccessorDecision.getData() == null) {
+                    throw ExceptionUtils.newException(InternalServerException.class, "필드 접근권한 조회 실패. type=%s, field=%s, result=%s", declaringClass, fieldName,
+                            resultFieldAccessorDecision);
                 }
-                TwoValueObject<Boolean, Integer> r = res.getData();
-                accessible = r.first;
-                handle = r.second;
+                FieldAccessAuthorityDecision fad = resultFieldAccessorDecision.getData();
+                accessible = fad.accessible;
+                handleType = fad.handleType;
+                handleBean = fad.handleBean;
             } else {
                 // handleType이 정해져 있으면 접근은 허용으로 간주 (필요 시 정책 조정)
-                accessible = false; // → false로 두고 'handle' 에 따라 변환 수행
+                accessible = false;
             }
         } catch (Exception e) {
             throw ExceptionUtils.newException(InternalServerException.class, e, "필드 처리결정 중 오류. type=%s, field=%s, 원인=%s", declaringClass, fieldName, e.getMessage());
         }
-
-        return new TwoValueObject<>(accessible, handle);
+        return new FieldAccessAuthorityDecision(accessible, handleType, handleBean);
     }
 }